ai_core/cognitive_seed/requirements.txt
numpy==1.26.4
scipy==1.13.0
scikit-learn==1.5.0
torch==2.3.0
pydantic==2.7.1
PyYAML==6.0.1
joblib==1.4.2
protobuf==5.26.1
cryptography==43.0.1
pyzmq==26.0.3
boto3==1.35.10
cbor2==5.6.2
redis==5.0.8
pyotp==2.9.0
argon2-cffi==23.1.0
python-dotenv==1.0.1